+++ Libplanet.Net/ActionExecutionState.cs
@@ -1,3 +1,4 @@
+using System;
 using Libplanet.Blocks;
 
 namespace Libplanet.Net
@@ -5,8 +6,7 @@ namespace Libplanet.Net
     /// <summary>
     /// Indicates a progress of executing block actions.
     /// </summary>
-    [Equals]
-    public class ActionExecutionState : PreloadState
+    public class ActionExecutionState : PreloadState, IEquatable<ActionExecutionState>
     {
         /// <summary>
         /// Total number of blocks to execute in the current batch.
@@ -27,9 +27,36 @@ namespace Libplanet.Net
         public override int CurrentPhase => 5;
 
         public static bool operator ==(ActionExecutionState left, ActionExecutionState right) =>
-            Operator.Weave(left, right);
+            left.Equals(right);
 
         public static bool operator !=(ActionExecutionState left, ActionExecutionState right) =>
-            Operator.Weave(left, right);
+            !left.Equals(right);
+
+        public bool Equals(ActionExecutionState? other)
+        {
+            if (other is null)
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, other))
+            {
+                return true;
+            }
+
+            return base.Equals(other) &&
+                   TotalBlockCount == other.TotalBlockCount &&
+                   ExecutedBlockCount == other.ExecutedBlockCount &&
+                   ExecutedBlockHash.Equals(other.ExecutedBlockHash);
+        }
+
+        public override bool Equals(object? obj) =>
+            obj is ActionExecutionState other && Equals(other);
+
+        public override int GetHashCode() => HashCode.Combine(
+                base.GetHashCode(),
+                TotalBlockCount,
+                ExecutedBlockCount,
+                ExecutedBlockHash);
     }
 }