+++ Libplanet.Net/BlockDownloadState.cs
@@ -1,3 +1,4 @@
+using System;
 using Libplanet.Blocks;
 
 namespace Libplanet.Net
@@ -5,8 +6,7 @@ namespace Libplanet.Net
     /// <summary>
     /// Indicates a progress of downloading blocks.
     /// </summary>
-    [Equals]
-    public class BlockDownloadState : PreloadState
+    public class BlockDownloadState : PreloadState, IEquatable<BlockDownloadState>
     {
         /// <summary>
         /// Total number of blocks to receive in the current batch.
@@ -32,9 +32,39 @@ namespace Libplanet.Net
         public BoundPeer? SourcePeer { get; internal set; }
 
         public static bool operator ==(BlockDownloadState left, BlockDownloadState right) =>
-            Operator.Weave(left, right);
+            left.Equals(right);
 
         public static bool operator !=(BlockDownloadState left, BlockDownloadState right) =>
-            Operator.Weave(left, right);
+            !left.Equals(right);
+
+        public bool Equals(BlockDownloadState? other)
+        {
+            if (other is null)
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, other))
+            {
+                return true;
+            }
+
+            return base.Equals(other) &&
+                   TotalBlockCount == other.TotalBlockCount &&
+                   ReceivedBlockCount == other.ReceivedBlockCount &&
+                   ReceivedBlockHash.Equals(other.ReceivedBlockHash) &&
+                   ((SourcePeer is null && other.SourcePeer is null) ||
+                    (SourcePeer is { } p1 && other.SourcePeer is { } p2 && p1.Equals(p2)));
+        }
+
+        public override bool Equals(object? obj) =>
+            obj is BlockDownloadState other && Equals(other);
+
+        public override int GetHashCode() => HashCode.Combine(
+                base.GetHashCode(),
+                TotalBlockCount,
+                ReceivedBlockCount,
+                ReceivedBlockHash,
+                SourcePeer);
     }
 }