+++ Libplanet/Store/Trie/Nodes/ValueNode.cs
@@ -1,3 +1,4 @@
+using System;
 using Bencodex.Types;
 
 namespace Libplanet.Store.Trie.Nodes
@@ -5,8 +6,7 @@ namespace Libplanet.Store.Trie.Nodes
     /// <summary>
     /// Wrapper class.
     /// </summary>
-    [Equals]
-    internal class ValueNode : INode
+    internal class ValueNode : INode, IEquatable<ValueNode>
     {
         public ValueNode(IValue value)
         {
@@ -15,11 +15,26 @@ namespace Libplanet.Store.Trie.Nodes
 
         public IValue Value { get; }
 
-        public static bool operator ==(ValueNode left, ValueNode right) =>
-            Operator.Weave(left, right);
+        public static bool operator ==(ValueNode left, ValueNode right) => left.Equals(right);
 
-        public static bool operator !=(ValueNode left, ValueNode right) =>
-            Operator.Weave(left, right);
+        public static bool operator !=(ValueNode left, ValueNode right) => !left.Equals(right);
+
+        public bool Equals(ValueNode? other)
+        {
+            if (other is null)
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, other))
+            {
+                return true;
+            }
+
+            return Value.Equals(other.Value);
+        }
+
+        public override bool Equals(object? obj) => obj is ValueNode other && Equals(other);
 
         /// <inheritdoc cref="INode.ToBencodex()"/>
         public IValue ToBencodex() =>